+++ libraries/exoplayer_ima/src/main/java/androidx/media3/exoplayer/ima/AdTagLoader.java
@@ -29,7 +29,6 @@ import static java.lang.annotation.ElementType.TYPE_USE;
 import android.content.Context;
 import android.os.Handler;
 import android.os.SystemClock;
-import android.text.TextUtils;
 import android.view.ViewGroup;
 import androidx.annotation.IntDef;
 import androidx.annotation.Nullable;
@@ -47,7 +46,6 @@ import androidx.media3.common.util.Util;
 import androidx.media3.datasource.DataSpec;
 import androidx.media3.exoplayer.source.ads.AdsLoader.EventListener;
 import androidx.media3.exoplayer.source.ads.AdsMediaSource.AdLoadException;
-import com.google.ads.interactivemedia.v3.api.Ad;
 import com.google.ads.interactivemedia.v3.api.AdDisplayContainer;
 import com.google.ads.interactivemedia.v3.api.AdError;
 import com.google.ads.interactivemedia.v3.api.AdErrorEvent;
@@ -75,7 +73,6 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -181,9 +178,6 @@ import java.util.Objects;
   /** Whether IMA has been notified that playback of content has finished. */
   private boolean sentContentComplete;
 
-  /** Mimetypes by AdInfo. Provided in the {@link AdEventType#LOADED} event */
-  private final Map<AdInfo, String> mimeTypeByAdInfo;
-
   // Fields tracking the player/loader state.
 
   /** Whether the player is playing an ad. */
@@ -254,7 +248,6 @@ import java.util.Objects;
     }
     imaSdkSettings.setPlayerType(IMA_SDK_SETTINGS_PLAYER_TYPE);
     imaSdkSettings.setPlayerVersion(IMA_SDK_SETTINGS_PLAYER_VERSION);
-    mimeTypeByAdInfo = new HashMap<>();
     this.supportedMimeTypes = supportedMimeTypes;
     this.adTagDataSpec = adTagDataSpec;
     this.adsId = adsId;
@@ -792,16 +785,6 @@ import java.util.Objects;
         String message = "AdEvent: " + adData;
         Log.i(TAG, message);
         break;
-      case LOADED:
-        @Nullable Ad ad = adEvent.getAd();
-        if (ad != null) {
-          String adMimeType = ad.getContentType();
-          if (!TextUtils.isEmpty(adMimeType)) {
-            // TODO: b/437870080 - Remove this once the SDK provides a mime type in the AdPodInfo.
-            setAdMimeType(ad.getAdPodInfo(), adMimeType);
-          }
-        }
-        break;
       default:
         break;
     }
@@ -1008,9 +991,16 @@ import java.util.Objects;
     }
 
     MediaItem.Builder adMediaItem = new MediaItem.Builder().setUri(adMediaInfo.getUrl());
-    String adMimeType = mimeTypeByAdInfo.get(adInfo);
-    if (adMimeType != null) {
-      adMediaItem.setMimeType(adMimeType);
+    // Use the video MIME type if it is provided.
+    // Demuxed streams may contain an audio MIME type, however it should only be used to set the
+    // audio MIME type or compose the audio codec string, when/if ExoPlayer introduces support for
+    // demuxed streams functionality. Even audio-only streams should only use the video MIME type as
+    // they are not demuxed. It is possible that the video MIME type is not provided, in which case,
+    // we do not set the MIME type of the MediaItem. However, if an audio MIME type is provided, it
+    // is most likely that the video MIME type is also provided (though not the other way around).
+    String videoMimeType = adMediaInfo.getVideoMimeType();
+    if (videoMimeType != null) {
+      adMediaItem.setMimeType(videoMimeType);
     }
 
     adPlaybackState =
@@ -1280,13 +1270,6 @@ import java.util.Objects;
               AdLoadException.createForUnexpected(new RuntimeException(message, cause)),
               adTagDataSpec);
     }
-  }
-
-  private void setAdMimeType(AdPodInfo adPodInfo, String mimeType) {
-    int adGroupIndex = getAdGroupIndexForAdPod(adPodInfo);
-    int adIndexInAdGroup = adPodInfo.getAdPosition() - 1;
-    AdInfo adInfo = new AdInfo(adGroupIndex, adIndexInAdGroup);
-    mimeTypeByAdInfo.put(adInfo, mimeType);
   }
 
   private int getAdGroupIndexForAdPod(AdPodInfo adPodInfo) {