+++ libraries/exoplayer_ima/src/test/java/androidx/media3/exoplayer/ima/ImaAdsLoaderTest.java
@@ -113,7 +113,9 @@ public final class ImaAdsLoaderTest {
   private static final MediaItem TEST_MEDIA_ITEM = MediaItem.fromUri(TEST_URI);
   private static final DataSpec TEST_DATA_SPEC = new DataSpec(TEST_URI);
   private static final Object TEST_ADS_ID = new Object();
-  private static final AdMediaInfo TEST_AD_MEDIA_INFO = new AdMediaInfo("https://www.google.com");
+  private static final AdMediaInfo TEST_AD_MEDIA_INFO =
+      new AdMediaInfo(
+          "https://www.google.com", /* audioMimeType= */ null, /* videoMimeType= */ null);
   private static final long TEST_AD_DURATION_US = 5 * C.MICROS_PER_SECOND;
   private static final ImmutableList<Float> PREROLL_CUE_POINTS_SECONDS = ImmutableList.of(0f);
 
@@ -292,17 +294,18 @@ public final class ImaAdsLoaderTest {
   }
 
   @Test
-  public void loadAd_withAdContentTypeSet_setsMimeTypeInAdPlaybackState() {
-    // Load the preroll ad with content type set. Intentionally use all lower-case HLS MIME type as
-    // this is what the IMA SDK sets.
-    when(mockPrerollSingleAd.getContentType()).thenReturn("application/x-mpegurl");
+  public void loadAd_withMimeType_setMimeTypeInAdPlaybackState() {
     imaAdsLoader.start(
         adsMediaSource, TEST_DATA_SPEC, TEST_ADS_ID, adViewProvider, adsLoaderListener);
 
+    videoAdPlayer.loadAd(
+        new AdMediaInfo(
+            "https://www.google.com", MimeTypes.APPLICATION_M3U8, MimeTypes.APPLICATION_M3U8),
+        mockAdPodInfo);
     adEventListener.onAdEvent(getAdEvent(AdEventType.LOADED, mockPrerollSingleAd));
-    videoAdPlayer.loadAd(TEST_AD_MEDIA_INFO, mockAdPodInfo);
 
-    // Verify that the preroll ad has been marked with the expected MIME type.
+    // Verify that the preroll ad has been marked with the expected MIME type. Provided in the ad
+    // media info.
     assertThat(getAdPlaybackState(/* periodIndex= */ 0))
         .isEqualTo(
             new AdPlaybackState(TEST_ADS_ID, /* adGroupTimesUs...= */ 0)
@@ -321,20 +324,11 @@ public final class ImaAdsLoaderTest {
     imaAdsLoader.start(
         adsMediaSource, TEST_DATA_SPEC, TEST_ADS_ID, adViewProvider, adsLoaderListener);
 
-    Ad mockFirstAd = mock(Ad.class);
-    AdPodInfo mockFirstAdPodInfo = mock(AdPodInfo.class);
-    when(mockFirstAdPodInfo.getPodIndex()).thenReturn(1);
-    when(mockFirstAdPodInfo.getTotalAds()).thenReturn(1);
-    when(mockFirstAdPodInfo.getAdPosition()).thenReturn(2);
-    when(mockFirstAd.getAdPodInfo()).thenReturn(mockFirstAdPodInfo);
-    when(mockFirstAd.getContentType()).thenReturn("application/x-mpegurl");
-
-    adEventListener.onAdEvent(getAdEvent(AdEventType.LOADED, mockFirstAd));
     videoAdPlayer.loadAd(TEST_AD_MEDIA_INFO, mockAdPodInfo);
     adEventListener.onAdEvent(getAdEvent(AdEventType.LOADED, mockPrerollSingleAd));
 
-    // Verify that the preroll ad has not been marked with the MIME type provided in the delayed
-    // LOADED event.
+    // Verify that the preroll ad has not been marked with the MIME type. As it is not provided in
+    // the ad media info.
     assertThat(getAdPlaybackState(/* periodIndex= */ 0))
         .isEqualTo(
             new AdPlaybackState(TEST_ADS_ID, /* adGroupTimesUs...= */ 0)